refactor: extract schedule tool subcomponents
